# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Swagger API docs for the GoChat API.",
        "title": "GoChat Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/sessions": {
            "post": {
                "description": "传入参数，用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "传入参数，注册用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "注册用户",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/users/:id": {
            "get": {
                "description": "传入参数，获取用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "获取用户信息",
                "responses": {
                    "201": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUserInfoResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "鉴权失败",
                        "schema": {
                            "$ref": "#/definitions/common.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "description": "传入参数，获取当前用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "获取当前用户信息",
                "responses": {
                    "201": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUserInfoResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "鉴权失败",
                        "schema": {
                            "$ref": "#/definitions/common.UnauthorizedResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.BadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "参数错误"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "common.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "鉴权失败"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "dto.CreateUserRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://avatars.githubusercontent.com/u/123456?v=4"
                },
                "email": {
                    "type": "string",
                    "example": "robin@qq.com"
                },
                "nickname": {
                    "type": "string",
                    "example": "robin"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2,
                    "example": "robin"
                }
            }
        },
        "dto.CreateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CreateUserResponseData"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dto.CreateUserResponseData": {
            "type": "object",
            "properties": {
                "createAt": {
                    "description": "可能需要格式化为字符串",
                    "type": "string",
                    "example": "2025-11-23T15:53:56.811"
                },
                "username": {
                    "description": "注意: 响应中不包含 PasswordHash",
                    "type": "string",
                    "example": "robin"
                },
                "uuid": {
                    "type": "string",
                    "example": "db376853-8f93-41f9-9a44-3c5ad8eedbbb"
                }
            }
        },
        "dto.GetUserInfoData": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://avatars.githubusercontent.com/u/123456?v=4"
                },
                "email": {
                    "type": "string",
                    "example": "robin@test.com"
                },
                "nickname": {
                    "type": "string",
                    "example": "robin"
                },
                "username": {
                    "type": "string",
                    "example": "robin"
                },
                "uuid": {
                    "type": "string",
                    "example": "db376853-8f93-41f9-9a44-3c5ad8eedbbb"
                }
            }
        },
        "dto.GetUserInfoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.GetUserInfoData"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "robin"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.LoginResponseData"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dto.LoginResponseData": {
            "type": "object",
            "properties": {
                "expireAt": {
                    "type": "integer",
                    "example": 1763910483465
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnb2NoYXQtYXBpLXNlcnZpY2UiLCJleHAiOjE3NjM5OTc1MTksImlhdCI6MTc2MzkxMTExOSwidXNlcklkIjoiY2YzZDE1ZmMtM2ZlYS00NDkzLWFjMDMtYzBiMjkzYTBjNjc4IiwidXNlcm5hbWUiOiJyb2JpbjYifQ.Rddl8mxWIPBWYCIO5TQYTG8uvyyPbP3FF9ozGdfytwg"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}